requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.6
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
openai>=1.40.0
markdown>=3.6
itsdangerous>=2.1.0
